chore: move tests to isolated folder
